Experiment: PNPM + Wireit (#6)
- Adds Wireit to previous pnpm scenario. Closes #4 
- Adds a `build:watch` task using wireit watching.
- Adds back in `nps` (I know!!! 😂 ) so that we can easily define root `package.json:scripts` and execute them within the workspace packages. Workspace tasks are defined in `package-scripts.js` and root tasks are in ye olde root `package.json:scripts`.
- Add GH action caching

Here's a timing table of just the `build` step for the different scenarios. Note that they're pretty close to even without a cache, then Wireit wins for "no changes" and "change a file in a single package". Do note that if we changed a source file in the `core` package which other packages depend on then Wireit would have to rebuild everything.

| Experiment | Time | Notes |
| --- | --- | --- |
| pnpm | 7s | No cache |
| pnpm-wireit | 8s | |
| pnpm | 7s | Cached |
| pnpm-wireit | 2s |
| pnpm | 8s | Cached w/ single package source change |
| pnpm-wireit | 4s |